pyproject.toml
[tool.poetry]
name = "ai-test-generator"
version = "0.1.0"
description = "AI-powered test case generator for Python projects"
authors = ["Your Name <your.email@example.com>"]

[tool.poetry.dependencies]
python = "^3.8"
fire = "^0.5.0"
pyyaml = "^6.0"
requests = "^2.26.0"
openai = "^1.1.0"
anthropic = "^0.3.0"

[tool.poetry.dev-dependencies]
pytest = "^7.0.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
ai-test-generator = "ai_test_generator.cli:main"
